chore: 優化 data.py
1. mysql 創一個新使用者 for 本專案
2. 刪除不必要、意義不明的 columns
